Split modals into separate component
